now handles Open3.capture3 (and probably the rest of that module)
* ENHANCEMENT: spawn is now mocked well enough that Open3.capture3 behaves as though spawn is not being monkey-patched when ShellMock is enabled (and therefore, spawn most definitely is being monkey-patched)

In order to do this, 2 things changed:
1. What's written to the CommandStub#writer IO pipe
2. What monkey-patched #spawn passes to #__un_shell_mocked_spawn

Before this change, we tried to serialize the entire command "signature" (env, command, options) to
the CommandStub#writer pipe. This was done in the future-proofing hope that the information would
somehow be useful on the other side. However, IO objects, which Open3 methods pass to spawn, can't
be Marshaled, so we had to stop. Fortunately, all the env & options matching is done before
Marshaling, in StubRegistry#stub_matching, so we don't need them to count the calls. If it later
turns out we want to see the full signature after calling, we can devise some way to satisfy both
demands.

The second change is much simpler. In order to output the correct output, we append an echo invocation
to our oneliner in #spawn, and pass __un_shell_mocked_spawn our options so it redirects correctly.