need to better name the specification methods
in ShellMock stubs, there are 2 types of methods, matchers and specifiers:
* matchers (command string, env, options) tell ShellMock when a stub has been invoked
* specifiers (output, exitstatus) let the command satisfy other code that depends on its invocation

I haven't liked the `and_` prefix for the specifiers, and I've been wanting a better prefix for a while.
`to_` is good in the abstract, but completely goes against the Ruby idiom that methods that start with `to_` are used to convert the receiver into a specific representation (`to_s`, `to_json`, etc..).
`will_` is a good prefix for specification.